Microservices/sentence-similarity/requirements.txt
flask_socketio
websocket-client
sentence-transformers
